moved ws injection
