+++ src/ws.ts
@@ -1,0 +1,13 @@
+let _WS: typeof WebSocket;
+
+if (typeof WebSocket !== 'undefined') {
+	_WS = WebSocket;
+}
+
+export function injectWebSocketImpl(ws: any) {
+	_WS = ws;
+}
+
+export function getWebSocketImpl() {
+	return _WS;
+}
