+++ test/wallet.test.ts
@@ -4,8 +4,8 @@ import { CashuWallet } from '../src/CashuWallet.js';
 import { CheckStateEnum, MeltQuoteResponse } from '../src/model/types/index.js';
 import { getDecodedToken } from '../src/utils.js';
 import { Proof } from '@cashu/crypto/modules/common';
-import { injectWebSocketImpl } from '../src/WSConnection.js';
 import { Server, WebSocket } from 'mock-socket';
+import { injectWebSocketImpl } from '../src/ws.js';
 
 injectWebSocketImpl(WebSocket);
